♻️ Put each error type in its own tracker file
